extract some classes to their own files
